% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaps-package.R
\docType{package}
\name{esaps-package}
\alias{esaps-package}
\title{Indicators of Electoral Systems and Party Systems}
\description{
It allows structuring electoral data of different size and structure
    to calculate various indicators frequently used in the studies of electoral
    systems and party systems:
        \itemize{

                \item{Electoral Volatility:}
                \itemize{
                        \item{\href{https://ejpr.onlinelibrary.wiley.com/doi/full/10.1111/j.1475-6765.1979.tb01267.x}{Pedersen (1979)} <doi:10.1111/j.1475-6765.1979.tb01267.x>}
                        \item{\href{https://www.cambridge.org/core/journals/british-journal-of-political-science/article/revisiting-electoral-volatility-in-post-communist-countries-new-data-new-results-and-new-approaches/46E14577C9EB765C7BED1A9EA657005F}{Powell and Tucker (2014)} <doi:10.1017/S0007123412000531>}
                        \item{\href{https://journals.sagepub.com/doi/abs/10.1177/1354068818795191}{Lago and Torcal (2019)}  <doi:10.1177/1354068818795191>}
                }
                \item{Effective Number of Parties:}
                \itemize{
                        \item{\href{https://journals.sagepub.com/doi/pdf/10.1177/001041407901200101}{Laakso and Taagepera (1979)} <doi:10.1177/001041407901200101>}
                }

                \item{Electoral disproportionality:}
                \itemize{
                        \item{Rae (1971, ISBN:9780300015171)}
                        \item{\href{https://www.cambridge.org/core/journals/british-journal-of-political-science/article/theoretical-limits-of-maximum-distortion-some-analytic-expressions-for-electoral-systems/77CE5C85BDDDE59036B7E91D81FBF87E}{Loosemore and Hanby (1971)} <doi:10.1017/S000712340000925X>}
                        \item{Lijphart (1986, ISBN:0875860648, 0875860745)}
                        \item{\href{https://oxford.universitypressscholarship.com/view/10.1093/acprof:oso/9780198273479.001.0001/acprof-9780198273479}{Lijphart} (1994) <doi:10.1093/acprof:oso/9780198273479.001.0001>}
                        \item{\href{https://www.sciencedirect.com/science/article/pii/026137949190025N?via\%3Dihub}{Cox and Shugart (1991)} <doi:10.1016/0261-3794(91)90025-N>}
                        \item{\href{https://www.sciencedirect.com/science/article/pii/026137949190004C?via\%3Dihub}{Gallagher (1991)} <doi:10.1016/0261-3794(91)90004-C>}
                        }

                \item{Party System Nationalization:}
                \itemize{
                        \item{\href{https://journals.sagepub.com/doi/abs/10.1177/13540688030092002}{Mainwaring and Jones (2003)} <doi:10.1177/13540688030092002>}
                        \item{\href{https://press.princeton.edu/books/ebook/9781400826377/the-formation-of-national-party-systems}{Chhibber and Kollman} (2004, ISBN:9781400826377)}
                        }
                        }
}
\seealso{
Useful link: \url{https://nicolas-schmidt.github.io/esaps/index.html}
}
\author{
Nicolas Schmidt \email{nschmidt@cienciassociales.edu.uy}
}
\keyword{Political}
\keyword{science}
