% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaps-package.R
\docType{package}
\name{esaps-package}
\alias{esaps-package}
\title{Indicators of Electoral Systems and Party Systems}
\description{
It allows structuring electoral data of different size and structure
    to calculate various indicators frequently used in the studies of electoral
    systems and party systems:
        \itemize{

                \item{ Electoral Volatility:}
                \itemize{
                        \item{\href{http://bit.ly/2p4cyZQ}{Pedersen (1979)} <doi:10.1111/j.1475-6765.1979.tb01267.x>}
                        \item{\href{http://bit.ly/2FpoeBm}{Powell and Tucker (2014)} <doi:10.1017/S0007123412000531>}
                        \item{\href{http://bit.ly/2FD83zE}{Torcal and Lago} (2015, ISBN:9788415260356)}
                }
                \item{Effective Number of Parties:}
                \itemize{
                        \item{\href{http://bit.ly/2oYU4dV}{Laakso and Taagepera (1979)} <doi:10.1177/001041407901200101>}
                }

                \item{Electoral disproportionality:}
                \itemize{
                        \item{Rae (1971, ISBN:9780300015171)}
                        \item{\href{https://bit.ly/2B1mIDH}{Loosemore and Hanby (1971)} <doi:10.1017/S000712340000925X>}
                       \item{Lijphart (1986, ISBN:0875860648, 0875860745)}
                        \item{\href{https://bit.ly/2McLshB}{Lijphart} (1994) <doi:10.1093/acprof:oso/9780198273479.001.0001>}
                        \item{\href{https://bit.ly/2M9Bttr}{Cox and Shugart (1991)} <doi:10.1016/0261-3794(91)90025-N>}
                        \item{\href{https://bit.ly/2M6CPoy}{Gallagher (1991)} <doi:10.1016/0261-3794(91)90004-C>}
                        }

                \item{Party System Nationalization:}
                \itemize{
                        \item{\href{https://bit.ly/2w9QgtO}{Mainwaring and Jones (2003)} <doi:10.1177/13540688030092002>}
                        \item{\href{https://bit.ly/2oj76SY}{Chhibber and Kollman} (2004, ISBN:9781400826377)}
                        }
                        }
}
\seealso{
Useful link: \url{https://github.com/Nicolas-Schmidt/esaps}
}
\author{
Nicolas Schmidt \email{nschmidt@cienciassociales.edu.uy}
}
\keyword{Political}
\keyword{science}
